requirements.txt
numpy
networkx
pyyaml
pandas
boto3